separate application and view helpers
